defaults split into save and load (removes issues with circular imports)
